requirements.txt
llama-cpp-python
rich
prompt-toolkit
pyyaml
gradio
